+++ platform/reworkd_platform/web/api/agent/tools/code.py
@@ -2,7 +2,6 @@ from fastapi.responses import StreamingResponse as FastAPIStreamingResponse
 from lanarky.responses import StreamingResponse
 from langchain import LLMChain
 
-from reworkd_platform.web.api.agent.model_settings import create_model
 from reworkd_platform.web.api.agent.tools.tool import Tool
 
 
@@ -18,11 +17,10 @@ class Code(Tool):
     ) -> FastAPIStreamingResponse:
         from reworkd_platform.web.api.agent.prompts import code_prompt
 
-        llm = create_model(self.model_settings, streaming=True)
-        chain = LLMChain(llm=llm, prompt=code_prompt)
+        chain = LLMChain(llm=self.model, prompt=code_prompt)
 
         return StreamingResponse.from_chain(
             chain,
-            {"goal": goal, "language": self.model_settings.language, "task": task},
+            {"goal": goal, "language": self.language, "task": task},
             media_type="text/event-stream",
         )
