+++ platform/reworkd_platform/web/api/agent/tools/utils.py
@@ -4,9 +4,7 @@ from typing import List
 from fastapi.responses import StreamingResponse as FastAPIStreamingResponse
 from lanarky.responses import StreamingResponse
 from langchain import LLMChain
-
-from reworkd_platform.schemas import ModelSettings
-from reworkd_platform.web.api.agent.model_settings import create_model
+from langchain.chat_models.base import BaseChatModel
 
 
 @dataclass
@@ -17,20 +15,22 @@ class CitedSnippet:
 
 
 def summarize(
-    model_settings: ModelSettings, goal: str, query: str, snippets: List[CitedSnippet]
+    model: BaseChatModel,
+    language: str,
+    goal: str,
+    query: str,
+    snippets: List[CitedSnippet],
 ) -> FastAPIStreamingResponse:
     from reworkd_platform.web.api.agent.prompts import summarize_prompt
 
-    chain = LLMChain(
-        llm=create_model(model_settings, streaming=True), prompt=summarize_prompt
-    )
+    chain = LLMChain(llm=model, prompt=summarize_prompt)
 
     return StreamingResponse.from_chain(
         chain,
         {
             "goal": goal,
             "query": query,
-            "language": model_settings.language,
+            "language": language,
             "snippets": snippets,
         },
         media_type="text/event-stream",
