+++ platform/reworkd_platform/web/api/agent/tools/wikipedia_search.py
@@ -1,8 +1,7 @@
 from lanarky.responses import StreamingResponse
 from langchain import WikipediaAPIWrapper
 
-from reworkd_platform.schemas import ModelSettings
-from reworkd_platform.web.api.agent.tools.stream_mock import stream_string
+from reworkd_platform.web.api.agent.stream_mock import stream_string
 from reworkd_platform.web.api.agent.tools.tool import Tool
 
 
@@ -14,14 +13,12 @@ class Wikipedia(Tool):
     )
     public_description = "Search Wikipedia for historical information."
 
-    def __init__(self, model_settings: ModelSettings):
-        super().__init__(model_settings)
-        self.wikipedia = WikipediaAPIWrapper(
+    async def call(self, goal: str, task: str, input_str: str) -> StreamingResponse:
+        wikipedia_client = WikipediaAPIWrapper(
             wiki_client=None,  # Meta private value but mypy will complain its missing
         )
-
-    async def call(self, goal: str, task: str, input_str: str) -> StreamingResponse:
+        
         # TODO: Make the below async
-        wikipedia_search = self.wikipedia.run(input_str)
-        # return summarize(self.model_settings, goal, task, [wikipedia_search])
+        wikipedia_search = wikipedia_client.run(input_str)
+        # return summarize(self.model, self.language, goal, task, [wikipedia_search])
         return stream_string("Wikipedia is currently not working")
